+++ frontend/src/components/admin/tabs/flag-tab.vue
@@ -3,19 +3,37 @@
     <thead>
       <tr>
         <th>Description</th>
+        <th>Value</th>
         <th>Enabled</th>
       </tr>
     </thead>
     <tbody>
       <tr v-for="item in flags" :key="`f${item.key}`">
-        <td class="nameCol">
+        <td class="nameCol" :colspan="colspan(item)">
           <h4>{{ title(item) }}</h4>
           <p class="desc">
             {{ DESC[item.key] }}
           </p>
         </td>
+        <td v-if="item.key === 'BT'" class="bannerTextField">
+          <v-text-field
+            :model-value="banner"
+            label="Banner"
+            clearable
+            hide-details="auto"
+            @update:model-value="banner = $event"
+            @click:clear="banner = ''"
+          />
+        </td>
         <td>
+          <v-btn
+            v-if="item.key === 'BT'"
+            :icon="mdiContentSaveOutline"
+            title="Save Banner"
+            @click="changeCol(item.key, 'value', banner)"
+          />
           <v-checkbox
+            v-else
             :model-value="item.on"
             :true-value="true"
             :error-messages="getFormErrors(item.key, 'on')"
@@ -29,6 +47,7 @@
 </template>
 
 <script>
+import { mdiContentSaveOutline } from "@mdi/js";
 import { mapActions, mapState } from "pinia";
 
 import ADMIN_FLAGS from "@/choices/admin-flag-choices.json";
@@ -41,16 +60,33 @@ export default {
   components: {},
   data() {
     return {
+      DESC,
+      banner: "",
+      mdiContentSaveOutline,
       lastUpdate: {
         name: "",
         field: undefined,
       },
-      DESC,
     };
   },
   computed: {
     ...mapState(useCommonStore, ["formErrors"]),
-    ...mapState(useAdminStore, ["flags"]),
+    ...mapState(useAdminStore, {
+      flags: (state) => state.flags,
+    }),
+    storeBanner() {
+      for (const item of this.flags) {
+        if (item.key === "BT") {
+          return item.value;
+        }
+      }
+      return "";
+    },
+  },
+  watch: {
+    storeBanner(to) {
+      this.banner = to;
+    },
   },
   mounted() {
     this.loadTables(["Flag"]);
@@ -71,6 +107,9 @@ export default {
     title(item) {
       return ADMIN_FLAGS[item.key];
     },
+    colspan(item) {
+      return item.key === "BT" ? 1 : 2;
+    },
   },
 };
 </script>
@@ -89,5 +128,9 @@ export default {
   margin-top: 1em;
   margin-bottom: 0.5em;
   color: rgb(var(--v-theme-textSecondary));
+}
+
+.bannerTextField {
+  min-width: 16em;
 }
 </style>